Change `inertia_norm` from `RADIAL` to `BOTH`
The mode inertia in Ball & Gizon (2014) corresponds to `inertia_norm='BOTH'` rather than `'RADIAL'`. Local testing changed the example χ² in AMP2 from 1.393517... to 1.393483..., which is about the level I expected. The best-fitting mode frequencies differ by less than 5e-5. That's all the testing I did but is consistent with my expectations of it being a very small change.